install the azle extension with a new azle command
