+++ .github/workflows/test.yml
@@ -295,9 +295,6 @@ jobs:
               shell: bash -l {0}
               run: npm run lint
             - if: ${{ needs.release-candidate-deploy.outputs.should_run_tests }}
-              working-directory: ${{ matrix.example_directories }}
-              run: dfx start --clean --background --host 127.0.0.1:8000
-            - if: ${{ needs.release-candidate-deploy.outputs.should_run_tests }}
               shell: bash -l {0}
               working-directory: ${{ matrix.example_directories }}
               run: npm install
@@ -305,6 +302,12 @@ jobs:
               shell: bash -l {0}
               working-directory: ${{ matrix.example_directories }}
               run: npm link azle
+            - if: ${{ needs.release-candidate-deploy.outputs.should_run_tests }}
+              working-directory: ${{ matrix.example_directories }}
+              run: npx azle install-dfx-extension
+            - if: ${{ needs.release-candidate-deploy.outputs.should_run_tests }}
+              working-directory: ${{ matrix.example_directories }}
+              run: dfx start --clean --background --host 127.0.0.1:8000
               # This is to put wasmedge-quickjs in the correct location for the --native-compilation tests
             - if: ${{ needs.release-candidate-deploy.outputs.should_run_tests }}
               shell: bash -l {0}
