+++ src/compiler/index.ts
@@ -1,6 +1,4 @@
-import { IOType } from 'child_process';
 import { mkdirSync, writeFileSync } from 'fs';
-import { join } from 'path';
 
 import { compileRustCodeWithCandidAndCompilerInfo } from './compile_rust_code_with_candid_and_compiler_info';
 import { setupFileWatcher } from './file_watcher/setup_file_watcher';
@@ -12,7 +10,6 @@ import { prepareDockerImage } from './prepare_docker_image';
 import { prepareRustStagingArea } from './prepare_rust_staging_area';
 import { logSuccess, time, unwrap } from './utils';
 import { green } from './utils/colors';
-import { execSyncPretty } from './utils/exec_sync_pretty';
 import { GLOBAL_AZLE_CONFIG_DIR } from './utils/global_paths';
 import { CompilerInfo } from './utils/types';
 
@@ -43,8 +40,6 @@ async function azle() {
     if (commandExecuted === true) {
         return;
     }
-
-    installDfxExtension(stdioType);
 
     const {
         canisterName,
@@ -138,15 +133,4 @@ async function azle() {
 function createAzleDirectories() {
     mkdirSync(GLOBAL_AZLE_CONFIG_DIR, { recursive: true });
     mkdirSync('.azle', { recursive: true });
-}
-
-// TODO this is just temporary
-// TODO until we either make azle an official extension in the DFINITY dfx extensions repo
-// TODO or we have a better way for the developer to install the extension locally
-function installDfxExtension(stdioType: IOType) {
-    const dfxExtensionDirectoryPath = join(__dirname, '../../dfx_extension');
-    execSyncPretty(
-        `cd ${dfxExtensionDirectoryPath} && ./install.sh`,
-        stdioType
-    );
 }
